refactor(providers): move the managed voice-prep core beside its future caller
React/store/i18n-free by design (its own header says so); the S5 hook on
the kizuna-soniox descriptor becomes its only production caller.
